Hack to work around glibc's poor handing of //IGNORE
Glibc's iconv implementation does honor //IGNORE, but the result of
iconv() will always indicate failure when an illegal sequence is
encountered so we have to track if //IGNORE was requested manually and
ignore EILSEQ if it was. This is done by adding an extra field to the
end of the binary returned to track the IGNORE status (0 or 1).

This probably should be done differently (specify ignore as a flag when
opening the converter and track it that way) but it'll do for now.