Switch to using dynamically allocated output buffer instead of static one
Before, this port driver was unable to convert large inputs because the
output buffer was a global of a fixed size. The new approach is to
malloc a buffer the size of the input and then realloc if iconv returns
E2BIG. Tested on OSX and on OpenBSD. Have not tested with Linux's glibc
iconv yet.